fix(agent): build all platform binaries to the correct dist directory
The `make agent` target was outputting to dist/agent (wrong name, wrong
location). The download route at /api/agent/download serves from
apps/web/data/agent-dist/infrawatch-agent-{os}-{arch}, so the install
one-liner was always returning 503.

- Makefile: build linux/arm64, darwin/amd64/arm64, windows/amd64 as
  well as linux/amd64; output to apps/web/data/agent-dist/ with the
  correct naming convention; fix GOCACHE/GOPATH so the Docker build
  container can write its cache under --user
- .gitignore: exclude apps/web/data/ (generated binaries)
- heartbeat: extract readAllDisks() into disk_linux.go (build tag
  linux) and a stub disk_other.go (build tag !linux) so the agent
  cross-compiles cleanly for Windows/macOS
