+++ .gitignore
@@ -40,3 +40,6 @@ deploy/dev-tls/
 
 # Go build artifacts
 go.work.sum
+
+# Agent binaries served by the web app (generated by `make agent`)
+apps/web/data/
